Revert "Merge branch 'ue5-main' into ue4-main"
This reverts commit faf0e47db54f938b72b3aa74bbb3ae16cab9bd10, reversing
changes made to 5457b3cefbd75dfc8638ae89c7cdf6469eb8d80b.

+++ Source/CesiumRuntime/Private/CesiumRasterOverlay.cpp
@@ -118,11 +118,11 @@ void UCesiumRasterOverlay::Refresh() {
   this->AddToTileset();
 }
 
-double UCesiumRasterOverlay::GetMaximumScreenSpaceError() const {
+float UCesiumRasterOverlay::GetMaximumScreenSpaceError() const {
   return this->MaximumScreenSpaceError;
 }
 
-void UCesiumRasterOverlay::SetMaximumScreenSpaceError(double Value) {
+void UCesiumRasterOverlay::SetMaximumScreenSpaceError(float Value) {
   this->MaximumScreenSpaceError = Value;
   this->Refresh();
 }
