+++ Source/CesiumRuntime/Public/CesiumSunSky.h
@@ -85,7 +85,7 @@ protected:
       BlueprintReadWrite,
       Category = "Cesium|Date and Time",
       meta = (ClampMin = -12, ClampMax = 14))
-  double TimeZone = -5.0;
+  float TimeZone = -5.f;
 
   /**
    * The current solar time represented as hours from midnight.
@@ -98,7 +98,7 @@ protected:
       BlueprintReadWrite,
       Category = "Cesium|Date and Time",
       meta = (UIMin = 4, UIMax = 22, ClampMin = 0, ClampMax = 23.9999))
-  double SolarTime = 13.0;
+  float SolarTime = 13.f;
 
   /**
    * The day of the month.
@@ -150,7 +150,7 @@ protected:
       BlueprintReadWrite,
       Category = "Cesium|Date and Time",
       meta = (ClampMin = -360, ClampMax = 360))
-  double NorthOffset = -90.0;
+  float NorthOffset = -90.f;
 
   /**
    * Enables adjustment of the Solar Time for Daylight Saving Time (DST).
@@ -263,7 +263,7 @@ protected:
       BlueprintReadWrite,
       meta = (EditCondition = "UpdateAtmosphereAtRuntime"),
       Category = "Cesium|Atmosphere")
-  double InscribedGroundThreshold = 30.0;
+  float InscribedGroundThreshold = 30;
 
   /**
    * When the player pawn is above this height, which is expressed in kilometers
@@ -287,7 +287,7 @@ protected:
       BlueprintReadWrite,
       meta = (EditCondition = "UpdateAtmosphereAtRuntime"),
       Category = "Cesium|Atmosphere")
-  double CircumscribedGroundThreshold = 100.0;
+  float CircumscribedGroundThreshold = 100;
 
   /**
    * False: Use Directional Light component inside CesiumSunSky.
@@ -307,21 +307,21 @@ protected:
    * the Georeference origin.
    */
   UPROPERTY(EditDefaultsOnly, BlueprintReadOnly, Category = "Sun")
-  double Elevation = 0.f;
+  float Elevation = 0.f;
 
   /**
    * The current sun elevation, corrected for atmospheric diffraction, in
    * degrees above the horizontal, as viewed from the Georeference origin.
    */
   UPROPERTY(EditDefaultsOnly, BlueprintReadOnly, Category = "Sun")
-  double CorrectedElevation = 0.f;
+  float CorrectedElevation = 0.f;
 
   /**
    * The current sun azimuth in degrees clockwise from North toward East, as
    * viewed from the Georeference origin.
    */
   UPROPERTY(EditDefaultsOnly, BlueprintReadOnly, Category = "Sun")
-  double Azimuth = 0.f;
+  float Azimuth = 0.f;
 
   /**
    * A switch to toggle between desktop and mobile rendering code paths.
@@ -348,7 +348,7 @@ protected:
    * Default intensity of directional light that's spawned for mobile rendering.
    */
   UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = "Cesium|Mobile")
-  double MobileDirectionalLightIntensity = 6.f;
+  float MobileDirectionalLightIntensity = 6.f;
 
 public:
   UFUNCTION(
@@ -368,7 +368,7 @@ public:
    */
   UFUNCTION(BlueprintCallable, BlueprintPure, Category = Sun)
   static void GetHMSFromSolarTime(
-      double InSolarTime,
+      float InSolarTime,
       int32& Hour,
       int32& Minute,
       int32& Second);
@@ -406,8 +406,7 @@ protected:
    * @param Radius The radius in kilometers.
    */
   UFUNCTION(BlueprintCallable, Category = "Cesium")
-  void
-  SetSkyAtmosphereGroundRadius(USkyAtmosphereComponent* Sky, double Radius);
+  void SetSkyAtmosphereGroundRadius(USkyAtmosphereComponent* Sky, float Radius);
 
   /**
    * Update MobileSkySphere by calling its RefreshMaterial function
